+++ src/Console/ConsumerCommand.php
@@ -5,22 +5,14 @@ namespace Haltuf\RabbitMQ\Console;
 use Haltuf\RabbitMQ\Consumer\Consumer;
 use InvalidArgumentException;
 use Symfony\Component\Console\Attribute\AsCommand;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use UnexpectedValueException;
 
 #[AsCommand(name: 'rabbitmq:consumer', description: 'Run a RabbitMQ consumer')]
-final class ConsumerCommand extends Command
+final class ConsumerCommand extends BaseConsumerCommand
 {
-	/** @param array<string, Consumer> $consumers */
-	public function __construct(
-		private readonly array $consumers,
-	) {
-		parent::__construct();
-	}
-
 	protected function configure(): void
 	{
 		$this->addArgument('consumerName', InputArgument::REQUIRED, 'Name of the consumer');
@@ -46,18 +38,10 @@ final class ConsumerCommand extends Command
 			}
 		}
 
-		$this->consumers[$consumerName]->consume($secondsToLive);
-
-		return 0;
-	}
-
-	private function validateConsumer(string $name): void
-	{
-		if (!isset($this->consumers[$name])) {
-			throw new InvalidArgumentException(
-				"Consumer [$name] does not exist\n\n Available consumers: "
-				. implode('', array_map(static fn($s) => "\n\t- [{$s}]", array_keys($this->consumers))),
-			);
-		}
+		return $this->runConsumer(
+			$consumerName,
+			$output,
+			static fn(Consumer $consumer) => $consumer->consume($secondsToLive),
+		);
 	}
 }